refactor(rename): rename classes to more clearly express their intent
