--- lib/crc-model-collection.js
@@ -1,128 +1,0 @@
-'use strict';
-
-const requireDir = require('require-dir');
-const espree = require('espree');
-const estraverse = require('estraverse');
-const _ = require('lodash');
-const libCrc = requireDir('.', {
-    recurse: true,
-    camelcase: true
-});
-const CrcModel = libCrc.crcModel;
-const CrcModelFactory = libCrc.crcModelFactory;
-
-/**
- Generates a collection of CrcModels based on source code strings or buffers.
- */
-class CrcModelCollection {
-
-    /**
-     * Instantiate a new CrcModelCollection
-     *
-     * @param {string|buffer} code   The source code to be analyzed.
-     * @param {object} config espree configuration.
-     *
-     * @returns {CrcModelCollection} A collection of CrcModels and their relationships with one another.
-     */
-    constructor(code, config) {
-        this.code = _.toString(code);
-        this.config = config || libCrc.astConfig;
-        this.ast = espree.parse(this.code, this.config);
-        this.models = this.groupBy('name', CrcModelFactory.isDeclaration);
-        this.associateCollaborators();
-    }
-
-    /**
-     * Traverses the AST to identifier collaborators and associate them the objects that depend on them.
-     *
-     * @returns {void}
-     */
-    associateCollaborators() {
-        let getModel = (node) => {
-            if (!_.isNil(node) && _.has(node, 'name')) {
-                return this.find({name: node.name});
-            }
-            return (void 0);
-        };
-        estraverse.traverse(this.ast, {
-            enter: (node) => {
-                let model = getModel(node);
-                if (model && !CrcModelFactory.inOwnDeclarationScope(node, model)) {
-                    CrcModelFactory.getModelDeclaredInRange(node, this);
-                }
-            }
-        });
-    }
-
-    /**
-     * Traverses the ast, returning an array of all elements predicate returns truthy for.
-     *
-     * @param {function|object} predicate The function or identity invoked per iteration.
-     *
-     * @returns {array} Returns a new, filtered array of zero or more CrcModels.
-     */
-    filterBy(predicate) {
-        let nodes = [];
-        estraverse.traverse(this.ast, {
-            enter: (node) => {
-                if (!_.isNil(node) && predicate(node, this)) {
-                    nodes.push(CrcModelFactory.getCrcModelOptions(node));
-                }
-            }
-        });
-        return nodes;
-    }
-
-    /**
-     * Iterates over the CrcModels collection, returning the first element predicate returns truthy for.
-     *
-     * @param {function|object} predicate The function or object literal "matcher" invoked per iteration.
-     *
-     * @example
-     * let alpha = crcModelCollection.find({name: 'Alpha'});
-     * // => [object Object] // Alpha reference
-     *
-     * @returns {array} Returns the matched CrcModel, else undefined.
-     */
-    find(predicate) {
-        return _.find(this.models, predicate);
-    }
-
-  // TODO: This method is not yet implemented.
-  // REVIEW: This smells weird here. The CrcModelCollection is becoming a factory aggregator.
-    // getPrototypeOf(name) {
-    //     let prototypePredicate, protos, nodes;
-    //     prototypePredicate = (node) => 'Identifier' === node.type && 'prototype' === node.name;
-    //     protos = this.filterBy(prototypePredicate);
-    //     nodes = _.filter(protos, {
-    //         name: name
-    //     });
-    //     return nodes;
-    // }
-
-    /**
-     * Creates an object composed transformed AST nodes, each of which has
-     * property <code>name</code>. The <code>name</code> acts as "key" that can
-     * be used by other methods like {@link CrcModelCollection#find}. The order
-     * of grouped values is determined by the order they occur in the AST.
-     *
-     * @param {string} property  The property/key to group by.
-     * @param {function|object} predicate The iteratee to transform keys.
-     *
-     * @returns {array} Returns the composed aggregate object array.
-     */
-    groupBy(property, predicate) {
-        let options, models = [];
-        options = _.values(_.omitBy(this.filterBy(predicate), _.isNil));
-        _.chain(options)
-          .groupBy(property)
-          .map((opts) => {
-              models.push(CrcModel.factory(opts, this.ast));
-          })
-          .value();
-        return models;
-    }
-
-}
-
-module.exports = CrcModelCollection;
