+++ lib/crc-model-visitor.js
@@ -35,10 +35,10 @@ class CrcModelVisitor {
         return options;
     }
 
-    static getModelDeclaredInRange(node, crcModelCollection) {
+    static getModelDeclaredInRange(node, crcModelList) {
         let crc = null;
         let declarations = {};
-        _.map(crcModelCollection.models, (model) => {
+        _.map(crcModelList.models, (model) => {
             declarations[model.name] = _.pick(model.declaration, [
                 'start',
                 'end'
@@ -50,7 +50,7 @@ class CrcModelVisitor {
             });
             crc = null;
             if (isDeclaredRange) {
-                crc = crcModelCollection.find({
+                crc = crcModelList.find({
                     name
                 });
                 crc.collaborators.push(node);
